+++ compatibility_spark-3/src/test/resources/delta-cdc-to-snapshot/first-input.csv
